UUFV-438 Move getSelectOptionLabel
