--- libs/application/templates/parental-leave/src/lib/parentalLeaveClientUtils.ts
@@ -1,12 +1,0 @@
-type SelectOption = {
-  label: string
-  value: string
-}
-
-export const getSelectOptionLabel = (options: SelectOption[], id?: string) => {
-  if (id === undefined) {
-    return undefined
-  }
-
-  return options.find((option) => option.value === id)?.label
-}
